If surrounding the mv

+++ stuff/install.sh
@@ -112,7 +112,11 @@ location ~ /\.ht {
 }
 EOF
 	sudo mv default.conf /etc/nginx/conf.d/
-	sudo mv /etc/nginx/sites-enabled/default /etc/nginx/sites-disabled
+	
+	if [ -f /etc/nginx/sites-enabled/default ]; then
+		sudo mv /etc/nginx/sites-enabled/default /etc/nginx/sites-disabled
+	fi
+	
 	sudo /etc/init.d/nginx restart
 fi
